recommend preRetrieve

+++ kato-core/kato-redis/src/main/java/com/kato/pro/redis/RedisService.java
@@ -177,7 +177,7 @@ public class RedisService {
 	 * @param time  时间(秒) time要大于0 如果time小于等于0 将设置无限期
 	 * @return true成功 false 失败
 	 */
-	public Boolean set(String key, String value, Long time) {
+	public Boolean set(String key, String value, Integer time) {
 		try {
 			if (time > 0) {
 				redisTemplate.opsForValue().set(key, value, time, TimeUnit.SECONDS);
